requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0
django-cors-headers==3.14.0
drf-nested-routers==0.93.4
Celery==5.2.7
redis==4.5.4
pytest==7.3.1
pytest-django==4.5.2
factory-boy==3.2.1
Faker==18.10.0
black==23.3.0
flake8==6.0.0
isort==5.12.0
